telebot/requirements.txt
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
idna==2.10
loguru==0.5.3
pyTelegramBotAPI==3.7.3
python-dotenv==0.15.0
requests==2.25.0
six==1.15.0
urllib3==1.26.2
win32-setctime==1.0.3
